+++ modules/test-utils/src/null-device/null-device.ts
@@ -59,6 +59,10 @@ export class NullDevice extends Device {
   readonly canvasContext: NullCanvasContext;
   readonly lost: Promise<{reason: 'destroyed'; message: string}>;
 
+  static attach(handle: null): NullDevice {
+    return new NullDevice({});
+  }
+
   static async create(props: DeviceProps = {}): Promise<NullDevice> {
     // Wait for page to load: if canvas is a string we need to query the DOM for the canvas element.
     // We only wait when props.canvas is string to avoids setting the global page onload callback unless necessary.